# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Produtos",
        "description": "Documentação da API de Produtos usando Swagger(antes de criar os produtos a(s) categoria precisam ser criada primeiro)",
        "contact": {
            "email": "suporte@exemplo.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "Usúarios"
                ],
                "summary": "Listar todos os usuários",
                "description": "Retorna a lista completa de usúarios",
                "operationId": "getUsers",
                "responses": {
                    "200": {
                        "description": "Lista de usúarios retornada com sucesso"
                    }
                }
            },
            "post": {
                "tags": [
                    "Usúarios"
                ],
                "summary": "novo usúario",
                "description": "Criação do usúario",
                "operationId": "Criar um novo Usúario",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "João silva"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "2i2ww@google.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "T@rt@14#"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Mostra o usuário criado com sucesso"
                    }
                }
            }
        },
        "/api/users/{user}": {
            "get": {
                "tags": [
                    "Usúarios"
                ],
                "summary": "detalhes do usuário",
                "description": "Detalhamento do usúario",
                "operationId": "recuperar o usúario por id",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID do usúario (id e do tipo uuid e necessario copiar o codigo id gerado na lista de usúarios entre as aspas duplas e colar no campo abaixo)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostra os detalhes do usúario com sucesso"
                    }
                }
            },
            "put": {
                "tags": [
                    "Usúarios"
                ],
                "summary": "atualizar usúario",
                "description": "Atualização do usúario",
                "operationId": "Atualizar um usúario",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID do usúario (id e do tipo uuid e necessario copiar o codigo id gerado na lista de usúarios entre as aspas duplas e colar no campo abaixo)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "João silva souza"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "joao@google.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "T@rt@14#"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mostra atualização do usúario com sucesso"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Usúarios"
                ],
                "summary": "deletar usúario",
                "description": "categoria deletada",
                "operationId": "Deletar um usúario",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID do usúario (id e do tipo uuid e necessario copiar o codigo id gerado na lista de usúarios entre as aspas duplas e colar no campo abaixo)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostra o usúario deletado com sucesso"
                    }
                }
            }
        }
    },
...
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}